qemu: TSAN Clean up
- Fix 3 thread races detected by tsan
- Change apicbase to 64 bit variable to reflect what it holds

Message-Id: <20251211231155.1171717-2-marcmorcos@google.com>

+++ include/hw/i386/apic_internal.h
@@ -158,7 +158,7 @@ struct APICCommonState {
 
     MemoryRegion io_memory;
     X86CPU *cpu;
-    uint32_t apicbase;
+    uint64_t apicbase; /* All accesses to apicbase must use qatomic helpers. */
     uint8_t id; /* legacy APIC ID */
     uint32_t initial_apic_id;
     uint8_t version;
